Remove xpyb support.
It's Python 2 only and was never enabled for py2/3 on debian/ubuntu
so this should not change much.

Handle instantiating of XCBSurface like is currently done for
XlibSurface. Getting an instance through the C-API still works.

If support gets added back we should look into xcffib, which is
what cairocffi uses. It also has the advandage that we don't
have to link against it and can use it at runtime.