cairo.Error: add a status attribute; add CairoError alias
This extends the cairo.Error exception type to have a status attribute
containing the cairo_status_t constants. This is what cairocffi
implements and will also make testing for specific error cases easier.

Also adds an cairo.CairoError alias to be compatible with the default
cairocffi naming.

Switch some functions which raise cairo.Error on programming errors
to ValueError, so we have a valid status value in all cases.