surface: Tie lifetime of surface base objects to the surface, not the wrapper. Fixes #11
In case the cairo.Surface got deallocated it took the base object with it, but the surface
could still be referenced from some other place like a cairo.Context/SurfacePattern.
In case the referencing object would then try to use the surface things crash.

Tie the lifetime to the actual object instead using the user_data API.