Convert some manual tests to pytest tests and fix endianness issues.
This converts some manual tests using the stdlib array module and
numpy arrays in combination with image surfaces.

While this does not add any means to verify the result it tests the code
paths and the output can be inspected by commenting out the parts
deleting the temporary files.

This also tries to make these tests give the same result on big endian
machines. Special handling is needed since FORMAT_ARGB32 is native endian
as was pointed out in issue #23