Fix leak when using ImageSurface.get_data() with Python 3. Fixes #41
Turns out the buffer/memoryview API can only be used in implementations
of the buffer protocol of some object and the current approach
worked by accident and for some unknown reason leaked the exporter as
well.

This creates a new Python class implementing the buffer interface,
which takes the buffer information and is only used to keep the
image surface alive and to hand out memory views.

In case the last view dies, the proxy will die and unref the image
surface.