raster pattern: don't pass the pattern to the callback. See #48
The pattern argument in the C API might be some internal thing
and not a bug as suspected. Instead of trying to "fix" it by passing
the pattern where the callback was registered at, pass nothing instead.

https://bugs.freedesktop.org/show_bug.cgi?id=101866#c1

In case the pattern instance is needed in the callback it can easily
be passed by using Python closures.