Port from PyObject_AsWriteBuffer() to PyObject_GetBuffer()
PyObject_AsWriteBuffer() is removed in Python 3.10. Port things to use
PyObject_GetBuffer() and PyBuffer_Release() instead.

This does more or less the same, but instead of keeping the object which provides
the buffer alive we now keep the buffer itself alive, which in turn keeps a reference
on the exporting object.

See #190