Check that file objects take/return byte objects when passing them to pycairo
Otherwise they get used the first time in the cairo callback where all we
can do is set a cairo error, but can't pass the real error to the caller.

This tries to read/write a zero length bytes object when the object is
passed to pycairo, so we can raise an exception that's helpful to the user.

Fixes #157