docs: start using autodoc on the typing stubs
Instead of having docs in two places write the docs in the .pyi
file and pull them into sphinx via autodoc.

This means we have to use rst in the docstrings (unless we use some
sphinx plugin to default to another syntax) but at least the docs
are only in one place.

This only converts things for two small classes, just to show
how it would work.