Windows: fix DLL not found error in Python 3.8+
This upstreams a patch from gvsbuild. In Python 3.8, a new
os.add_dll_directory method was added to make loading DLLs in Windows
more consistent. The impact of this is that even if I compile GTK
successfully with MSVC, and add it to the Path, INCLUDE, and LIB, I am
still not able to pip install pycairo in Windows. This finds the first
GTK DLL location from the Windows PATH variable.