docs: add an example for how to convert a surface to pillow
In https://github.com/pygobject/pycairo/pull/280 the use case came up
to save the result as a jpeg, or other formats and allow more control
over the compression settings.

We don't really need to duplicate this logic with Pillow implementing
it, so add an example for how to convert an image surface to a Pillow
image object.