CI: disable build isolation and bump meson-python
Turns out we already depend on a newer version, as building with 0.12
fails here due some config keys being not optional there, so bump it.

To make it easier to test certain build tool versions add meson-python
as a dev dependency (matching the build-system version requirement, so
we are testing the same thing) and in CI build without build isolation,
so we are using meson-python as installed by poetry.