ImageSurface.get_data: raise in case the surface is finished + more docs
cairo_image_surface_get_data() is documented to return NULL in case the
surface is finished, but that is not the case. It just returns the potentially
freed pointer.

To avoid this being used raise if the surface is already finished. Since there
is no way to check if a surface is finished directly create a dummy context
and query its status instead.

This is potentially a breaking change, there are three cases when called
after finish():

* In the common case of a normal image surface the returned data was
  freed and potentially garbage.
* In case the image surface was sourced from some external buffer
  the returned data was still valid when created in Python,
  though there are plans to change that:
  https://github.com/pygobject/pycairo/issues/392
  When created externally the data is potentially garbage.
* In case the image was from a png it really returned None for
  some reason, though we never documented that case.

Also while at it, copy more from the upstream docs regarding flush() and
mark_dirty().