Surface: release the create_for_data() input buffer on finish()
In case a surface is created via create_for_data() we get a buffer
from the input for passing to cairo. Once the surface is finished
though, cairo should no longer read/write to it and the buffer is no
longer really needed. Up until now we only released the buffer once the
surface was destroyed, which means after the Python wrapper was GCed,
which is hard to control from Python if one wants to release resources.

This changes the buffer release to also happen on finish().
Note that this only works if finish() is called from Python since we
don't notice otherwise.

This is also related to https://github.com/pygobject/pycairo/pull/400
which made get_data() raise in case the surface is finished, to avoid
returning freed data with this on top.

Fixes #392