Deprecate the num_glyphs parameter everywhere
This deprecates the num_glyphs parameter for:

* Context.glyph_extents
* Context.glyph_path
* Context.show_glyphs
* ScaledFont.glyph_extents

It could be used to limit the length of the passed sequence to be
considered, while being limited to the real length to the sequence,
and also defaulting to the real length of the sequence.

This is just confusing since no other APIs work that way in pycairo.
I never added them to the stubs, and they are also gone from the
HTML docs since that was switched to be generated from the stubs.

Let's make this more official by emitting a deprecation warning
when they are used at runtime. There are no plans to remove them
though.