build: add no-x11 option to disable X11 surface support
This avoids using any of the xlib and xcb APIs provided by cairo
as if cairo was not built with support for those.

This is to allow downstream to provide non-x cairo builds, and
a pycairo build that works with both cairo variants.

PR #421 is a more dynamic solution for this problem, but this is easier
and less risky for now.

See #420